редактирование информации о пользователе

+++ CoffeeHousesRedact.xaml.cs
@@ -34,23 +34,28 @@ namespace coffee_app
 
         private void Button_Click(object sender, RoutedEventArgs e)
         {
-            foreach (var i in coffee.CoffeeHouse)
+            if (name.SelectedItem != null)
             {
-                if (name.SelectedItem.ToString() == i.name)
+                foreach (var i in coffee.CoffeeHouse)
                 {
-                    text.Document.Blocks.Clear();
-                    nameC.Document.Blocks.Clear();
-                    adres.Document.Blocks.Clear();
+                    if (name.SelectedItem.ToString() == i.name)
+                    {
+                        text.Document.Blocks.Clear();
+                        nameC.Document.Blocks.Clear();
+                        adres.Document.Blocks.Clear();
 
-                    text.AppendText(i.description);
-                    nameC.AppendText(i.name);
-                    adres.AppendText(i.address);
+                        text.AppendText(i.description);
+                        nameC.AppendText(i.name);
+                        adres.AppendText(i.address);
 
-                    CoffeeHouse = i;
+                        CoffeeHouse = i;
+                    }
                 }
+                but_save.Visibility = Visibility.Visible;
+                but_delete.Visibility = Visibility.Visible;
             }
-            but_save.Visibility = Visibility.Visible;
-            but_delete.Visibility = Visibility.Visible;
+            else
+                MessageBox.Show("Элемент не выбран!");
         }
 
         private void Button_Click_1(object sender, RoutedEventArgs e)
@@ -60,56 +65,66 @@ namespace coffee_app
 
         private void but_save_Click_1(object sender, RoutedEventArgs e)
         {
-            string description = new TextRange(text.Document.ContentStart, text.Document.ContentEnd).Text;
-            string adresText = new TextRange(adres.Document.ContentStart, adres.Document.ContentEnd).Text;
-            string nameText = new TextRange(nameC.Document.ContentStart, nameC.Document.ContentEnd).Text;
-
-            Functions f = new Functions();
-            f.RegCoffeeHouse(CoffeeHouse.idGuid, nameText, description, adresText);
-
-            foreach (var i in coffee.CoffeeHouse)
+            if (name.SelectedItem != null)
             {
-                if (i.name == name.SelectedItem.ToString())
+                string description = new TextRange(text.Document.ContentStart, text.Document.ContentEnd).Text;
+                string adresText = new TextRange(adres.Document.ContentStart, adres.Document.ContentEnd).Text;
+                string nameText = new TextRange(nameC.Document.ContentStart, nameC.Document.ContentEnd).Text;
+
+                Functions f = new Functions();
+                f.RegCoffeeHouse(CoffeeHouse.idGuid, nameText, description, adresText);
+
+                foreach (var i in coffee.CoffeeHouse)
                 {
-                    i.name = nameText;
-                    i.description = description;
-                    i.address = adresText;
+                    if (i.name == name.SelectedItem.ToString())
+                    {
+                        i.name = nameText;
+                        i.description = description;
+                        i.address = adresText;
+                    }
                 }
+
+                but_save.Visibility = Visibility.Hidden;
+                but_delete.Visibility = Visibility.Hidden;
+                text.Document.Blocks.Clear();
+                adres.Document.Blocks.Clear();
+                nameC.Document.Blocks.Clear();
+
+                coffee.SaveChanges();
+                name.Items.Clear();
+                foreach (var i in coffee.CoffeeHouse)
+                    name.Items.Add(i.name);
             }
-
-            but_save.Visibility = Visibility.Hidden;
-            but_delete.Visibility = Visibility.Hidden;
-            text.Document.Blocks.Clear();
-            adres.Document.Blocks.Clear();
-            nameC.Document.Blocks.Clear();
-
-            coffee.SaveChanges();
-            name.Items.Clear();
-            foreach (var i in coffee.CoffeeHouse)
-                name.Items.Add(i.name);
+            else
+                MessageBox.Show("Элемент не выбран!");
         }
 
         private void Button_Click_2(object sender, RoutedEventArgs e)
         {
-            foreach (var i in coffee.CoffeeHouse)
+            if (name.SelectedItem != null)
             {
-                if (name.SelectedItem.ToString() == i.name)
+                foreach (var i in coffee.CoffeeHouse)
                 {
-                    CoffeeHouse = i;
+                    if (name.SelectedItem.ToString() == i.name)
+                    {
+                        CoffeeHouse = i;
+                    }
                 }
+                Functions f = new Functions();
+                f.DeleteCoffeeHouse(CoffeeHouse);
+
+                but_save.Visibility = Visibility.Hidden;
+                but_delete.Visibility = Visibility.Hidden;
+                text.Document.Blocks.Clear();
+                adres.Document.Blocks.Clear();
+                nameC.Document.Blocks.Clear();
+
+                name.Items.Clear();
+                foreach (var i in coffee.CoffeeHouse)
+                    name.Items.Add(i.name);
             }
-            Functions f = new Functions();
-            f.DeleteCoffeeHouse(CoffeeHouse);
-
-            but_save.Visibility = Visibility.Hidden;
-            but_delete.Visibility = Visibility.Hidden;
-            text.Document.Blocks.Clear();
-            adres.Document.Blocks.Clear(); 
-            nameC.Document.Blocks.Clear();
-
-            name.Items.Clear();
-            foreach (var i in coffee.CoffeeHouse)
-                name.Items.Add(i.name);
+            else
+                MessageBox.Show("Элемент не выбран!");
         }
 
         private void Button_Click_3(object sender, RoutedEventArgs e)
